Override ModelChangeMixin.audit_compare() so that we can fix how changes to MultiSelectField fields are detected.
